NIflow Class

NIflow Class

NIflow

Format

R6Class object.

Value

Object of R6Class and NIflow.

Methods

Documentation

For full documentation of each method follow the corresponding link.

initialize(name, inputs = list())

Create a new flow with the given name and inputs. Documented in NIflow.initialize.

name()

Returns the name of the flow. Documented in NIflow.name.

get_private()

Returns the private environment inside the class object. Documented in NIflow.get_private.

set_name(new_name)

Changes the name of the flow. Documented in NIflow.set_name.

get_inputs()

Returns the inputs of the flow. Documented in NIflow.get_inputs.

get_outputs()

Returns the outputs of the flow. Documented in NIflow.get_outputs.

get_required_inputs(outputs)

Returns the inputs required to compute given outputs. Documented in NIflow.get_required_inputs.

get_process(output)

Returns the function that computes the given output. Documented in NIflow.get_process.

get_dependencies()

Returns the list of package dependencies of the flow. Documented in NIflow.get_dependencies.

check_dependencies()

Checks if all flow dependencies are installed. Documented in NIflow.check_dependencies.

replace(output, with)

Replace a function that computes an output with another function. Documented in NIflow.replace.

add(what = NULL, inputs = NULL, output = NULL, ...)

Adds a function to the flow, that transforms the given inputs into the given output. Documented in NIflow.add.

execute(inputs = list(), desired_outputs = NULL, initialize_outputs = TRUE, ...)

Execute the flow to obtain some outputs, given input files. Documented in NIflow.execute.

run(...)

Just a wrapper for execute. Documented in NIflow.run.

log(level = c("DEBUG", "INFO", "WARNING", "ERROR"), message = '...')

Add a message to the flow log, including the level of the message and a timestamp. Documented in NIflow.log.

print_log(level = c("DEBUG", "WARNING", "INFO", "ERROR"))

Prints the log of the flow. Documented in NIflow.print_log.

errors()

Just prints the errors in the flow log. Documented in NIflow.errors.

warnings()

Just prints the warnings in the flow log. Documented in NIflow.warnings.

save_log(filename, level = c("DEBUG", "WARNING", "INFO", "ERROR"))

Saves a the flow log to file. Documented in NIflow.save_log.

graph()

Returns the graph of the flow. Documented in NIflow.graph.

plot()

Plots the graph of the flow. Documented in NIflow.plot.

memory_used()

Total memory used by the flow, including its internal variables. Documented in NIflow.memory_used.

save(path = tempdir(), file_prefix = self$name())

Save the flow to disk. Documented in NIflow.save.

load(filename)

Load a flow from disk. Documented in NIflow.load.

deep_clone()

Clone a flow. Documented in NIflow.deep_clone.

subset(outputs)

Take a subset of the flow, for given outputs. Documented in NIflow.subset.

to_package(path, package_name = self$name())

Creates a package with functions provided by the execution of the flow. Documented in NIflow.to_package.

Examples

NIflow$new()

#> NIflow object with:
#>   name:  
#>   inputs:  
#>   outputs:  
#> Currently using 624.29 kB of memory.